# bff-node/docs/swagger.yaml
openapi: 3.0.0
info:
  title: BFF 101estoques
  version: 1.0.0
  description: Backend For Frontend para orquestração dos microserviços do 101estoques
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/estoques/{id}:
    get:
      summary: Retorna um estoque específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes do estoque
    put:
      summary: Atualiza um estoque
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estoque atualizado
    delete:
      summary: Remove um estoque
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estoque removido
...